requirements.txt
pandas
requests
sqlalchemy
python-dotenv
psycopg2-binary
pendulum
jinja2
